requirements.txt
# Core data & visualization
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.17.0

# NLP & ML
nltk>=3.8.1
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
umap-learn>=0.5.0
bertopic>=0.15.0
wordcloud>=1.9.0

# Streaming web app
streamlit>=1.28.0

# Geographic data
geopandas>=0.13.0
pydeck>=0.8.0

# Cloud & API
google-cloud-bigquery>=3.14.0
google-auth>=2.25.0
requests>=2.31.0
openai>=1.0.0

# Country data
pycountry>=23.12.11
